✨ refactor: centralize organization types and improve imports
- Refactor organization-related components to import types from '@/types/intervention' for better maintainability.
- Remove redundant type definitions from organizationService to streamline the codebase.
- Add eslint-disable comments in useEffect hooks to address warnings related to dependencies.
