+++ src/app/dashboard/(registre-securite)/interventions/organizations/_components/OrganizationCard.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { Organization } from '@/services/organizationService';
+import { Organization } from '@/types/intervention';
 import {
   Delete as DeleteIcon,
   Edit as EditIcon
